+++ Projectiles/Ranged/OpalStrikerHoldout.cs
@@ -19,14 +19,16 @@ namespace CalamityMod.Projectiles.Ranged
         private Player Owner => Main.player[Projectile.owner];
         public SlotId OpalChargeSlot;
         public SlotId OpalChargeLoopSlot;
+        public SlotId OpalChargeShotSlot;
 
         private float CurrentChargingFrames = 0f;
 
         private bool OwnerCanShoot => Owner.channel && !Owner.noItems && !Owner.CCed;
         private float storedVelocity = 12f;
         public const float velocityMultiplier = 1.2f;
-        public bool Extradamage;
+        public bool Extradamage = false;
         public int Time = 0;
+        public int Aftershot = 30;
 
         public override string Texture => "CalamityMod/Items/Weapons/Ranged/OpalStriker";
 
@@ -49,25 +51,47 @@ namespace CalamityMod.Projectiles.Ranged
                 Extradamage = false;
 
             Vector2 armPosition = Owner.RotatedRelativePoint(Owner.MountedCenter, true);
-            Vector2 tipPosition = armPosition + Projectile.velocity * Projectile.width * 0.5f;
+            Vector2 tipPosition = armPosition + Projectile.velocity * Projectile.width * 0.85f;
 
             // Fire if the owner stops channeling or otherwise cannot use the weapon.
             if (!OwnerCanShoot)
             {
-                int newdamage =  Projectile.damage * (Extradamage ? 9 : 1);
-                SoundEngine.PlaySound(OpalStriker.Fire, Projectile.position);
-                CurrentChargingFrames = 0;
-                Vector2 shootVelocity = Projectile.velocity.SafeNormalize(Vector2.UnitY) * storedVelocity;
-                Projectile.NewProjectile(Projectile.GetSource_FromThis(), tipPosition, shootVelocity, ModContent.ProjectileType<OpalStrike>(), newdamage, Projectile.knockBack * (Extradamage ? 3 : 1), Projectile.owner);
-                Projectile.Kill();
+                if (Aftershot == 30)
+                {
+                    int newdamage = Projectile.damage * (Extradamage ? 5 : 1);
+                    OpalChargeShotSlot = SoundEngine.PlaySound(Extradamage ? OpalStriker.ChargedFire : OpalStriker.Fire, Projectile.position);
+                    CurrentChargingFrames = 0;
+                    Vector2 shootVelocity = Projectile.velocity.SafeNormalize(Vector2.UnitY) * storedVelocity;
+                    if (Extradamage)
+                    {
+                        Projectile.NewProjectile(Projectile.GetSource_FromThis(), tipPosition, shootVelocity, ModContent.ProjectileType<OpalChargedStrike>(), newdamage, Projectile.knockBack * 3, Projectile.owner);
+                        for (int i = 0; i <= 10; i++)
+                        {
+                            Dust dust = Dust.NewDustPerfect(tipPosition, 162, shootVelocity.RotatedByRandom(MathHelper.ToRadians(20f)) * Main.rand.NextFloat(0.8f, 1.4f), 0, default, Main.rand.NextFloat(1.5f, 2.3f));
+                            dust.noGravity = true;
+                        }
+                    }
+                    else
+                        Projectile.NewProjectile(Projectile.GetSource_FromThis(), tipPosition, shootVelocity, ModContent.ProjectileType<OpalStrike>(), newdamage, Projectile.knockBack, Projectile.owner);
+                    
+                    Aftershot--;
+                }
+                else
+                    Aftershot--;
+
+                if (Aftershot == 0)
+                    Projectile.Kill();
             }
             else
             {
                 Player player = Main.player[Projectile.owner];
                 if (CurrentChargingFrames >= 10)
                 {
-                    Particle streak = new ManaDrainStreak(player, Main.rand.NextFloat(0.06f + (CurrentChargingFrames / 180), 0.08f + (CurrentChargingFrames / 180)), Main.rand.NextVector2CircularEdge(2f, 2f) * Main.rand.NextFloat(0.3f * CurrentChargingFrames, 0.3f * CurrentChargingFrames), 0f, Color.Gold, Color.Orange, 7, tipPosition);
-                    GeneralParticleHandler.SpawnParticle(streak);
+                    if (CurrentChargingFrames < 88)
+                    {
+                        Particle streak = new ManaDrainStreak(player, Main.rand.NextFloat(0.06f + (CurrentChargingFrames / 180), 0.08f + (CurrentChargingFrames / 180)), Main.rand.NextVector2CircularEdge(2f, 2f) * Main.rand.NextFloat(0.3f * CurrentChargingFrames, 0.3f * CurrentChargingFrames), 0f, Color.Gold, Color.Orange, 7, tipPosition);
+                        GeneralParticleHandler.SpawnParticle(streak);
+                    }
                     Particle orb = new GenericBloom(tipPosition, Projectile.velocity, Color.Orange, CurrentChargingFrames / 135, 2);
                     GeneralParticleHandler.SpawnParticle(orb);
                     Particle orb2 = new GenericBloom(tipPosition, Projectile.velocity, Color.Khaki, CurrentChargingFrames / 200, 2);
@@ -90,7 +114,6 @@ namespace CalamityMod.Projectiles.Ranged
                     {
                         OpalChargeLoopSlot = SoundEngine.PlaySound(OpalStriker.ChargeLoop, Projectile.position);
                     }
-                    Item heldItem = Owner.ActiveItem();
                 }
 
                 if (CurrentChargingFrames == 88f)
@@ -98,8 +121,8 @@ namespace CalamityMod.Projectiles.Ranged
                     for (int i = 0; i < 36; i++)
                     {
                         Dust chargefull = Dust.NewDustPerfect(tipPosition, 162);
-                        chargefull.velocity = (MathHelper.TwoPi * i / 36f).ToRotationVector2() * 18f + Owner.velocity;
-                        chargefull.scale = Main.rand.NextFloat(1f, 1.5f);
+                        chargefull.velocity = (MathHelper.TwoPi * i / 36f).ToRotationVector2() * 13f + Owner.velocity;
+                        chargefull.scale = Main.rand.NextFloat(2f, 2.5f);
                         chargefull.noGravity = true;
                     }
                 }
@@ -108,6 +131,8 @@ namespace CalamityMod.Projectiles.Ranged
                 OpalChargeLoop.Position = Projectile.Center;
             if (SoundEngine.TryGetActiveSound(OpalChargeSlot, out var OpalCharge) && OpalCharge.IsPlaying)
                 OpalCharge.Position = Projectile.Center;
+            if (SoundEngine.TryGetActiveSound(OpalChargeShotSlot, out var OpalChargeShot) && OpalChargeShot.IsPlaying)
+                OpalChargeShot.Position = Projectile.Center;
             UpdateProjectileHeldVariables(armPosition);
             ManipulatePlayerVariables();
         }
@@ -121,7 +146,7 @@ namespace CalamityMod.Projectiles.Ranged
                 Projectile.velocity = Vector2.Lerp(Projectile.velocity, Projectile.SafeDirectionTo(Main.MouseWorld), interpolant);
             }
             Player player = Main.player[Projectile.owner];
-            Projectile.position = player.RotatedRelativePoint(player.MountedCenter, true) - Projectile.Size / 2f;
+            Projectile.Center = player.RotatedRelativePoint(player.MountedCenter, true) + Projectile.velocity * 15;
             Projectile.rotation = Projectile.velocity.ToRotation() + (Projectile.spriteDirection == -1 ? MathHelper.Pi : 0f);
             Projectile.spriteDirection = Projectile.direction;
             Projectile.timeLeft = 2;
